Delete "open" command and change "begin" to accept an arbitrary message ID.
Since writing the "open" command, I realized that anyone can get a
message ID by clicking (you guessed it) "Copy ID" in the menu for that
message. That means we can just pass in the ID of the message to check,
and can also delete a lot of global state. (Mapping guilds to message
IDs, for example, since we no longer need to track that..)